Implement code review changes on all files
