+++ src/components/AddBreakButton.jsx
@@ -1,17 +1,21 @@
-import React from "react";
-import { Button } from "antd";
-import { PlusOutlined } from "@ant-design/icons";
-import styled from "styled-components";
+import { PlusOutlined } from '@ant-design/icons';
+import { Button } from 'antd';
+import React from 'react';
+import styled from 'styled-components';
 
-const AddBreakButton = (props) => {
+function AddBreakButton(props) {
+  const { setChangedAddBreakButton } = props;
+
   return (
     <ParentContainer>
-      <ButtonAddBreak block onClick={props.setChangedAddBreakButton}>
-        Add Break <PlusOutlined />
+      <ButtonAddBreak block onClick={setChangedAddBreakButton}>
+        Add Break
+        {' '}
+        <PlusOutlined />
       </ButtonAddBreak>
     </ParentContainer>
   );
-};
+}
 
 const ParentContainer = styled.div`
   padding-bottom: 0.5em;
@@ -20,18 +24,18 @@ const ParentContainer = styled.div`
 const ButtonAddBreak = styled(Button)`
   background-color: #b8e3fd;
   border-color: #b8e3fd;
-  color: black;
+  color: #000;
   height: 64px;
   width: 100%;
 
   :hover {
     background-color: #73c8fb;
-    color: black;
+    color: #000;
   }
 
   :focus {
     background-color: #73c8fb;
-    color: black;
+    color: #000;
   }
 `;
