+++ src/components/BreakOption.jsx
@@ -1,18 +1,24 @@
-import { React, useState } from "react";
-import { Card, Row, Col, Button } from "antd";
-import styled from "styled-components";
+import {
+  Button, Card, Col, Row,
+} from 'antd';
+import { React } from 'react';
+import styled from 'styled-components';
 
-function BreakOption({
-  setChangedBreakDisplay,
-  isShortDisabled,
-  isLongDisabled,
-}) {
+import { SHORT_BREAK, LONG_BREAK } from '../constants/common';
+
+function BreakOption(props) {
+  const {
+    setChangedBreakDisplay,
+    isShortDisabled,
+    isLongDisabled,
+  } = props;
+
   function handleOnClickShortBreak() {
-    setChangedBreakDisplay("Short Break");
+    setChangedBreakDisplay(SHORT_BREAK);
   }
 
   function handleOnClickLongBreak() {
-    setChangedBreakDisplay("Long Break");
+    setChangedBreakDisplay(LONG_BREAK);
   }
 
   return (
@@ -43,6 +49,7 @@ function BreakOption({
     </ParentContainer>
   );
 }
+
 const ParentContainer = styled.div`
   padding-bottom: 0.5em;
 `;
@@ -56,14 +63,13 @@ const Column = styled(Col)`
   display: flex;
   justify-content: center;
   align-items: center;
-  /* color: white; */
 `;
 
 const BreakButton = styled(Button)`
   width: 100%;
   background-color: transparent;
-  color: black;
-  border-color: black;
+  color: #000;
+  border-color: #000;
 `;
 
 export default BreakOption;
