+++ src/constants/common.js
@@ -1,0 +1,9 @@
+export const SHORT_BREAK_MINUTES = 5;
+export const LONG_BREAK_MINUTES = 15;
+export const SHORT_BREAK = 'Short Break';
+export const LONG_BREAK = 'Long Break';
+export const TASK_TYPES = {
+  BREAK: 'break',
+  TASK: 'task',
+};
+export const DEFAULT_TASK_MINUTES = 25;
